:recycle: Refactor Reservations
The reservation entity now has a new atributte that shows if it's a reservation or a queue ticket.
Now there's no queue ticket entities.
The management of the queue is better 'cause it only needs one view instend of an union of queries.

TO DO: fact queue - the view that show the order of reservations - needs a DTO.